+++ src/source_analysis/attributes.rs
@@ -64,19 +64,41 @@ pub(crate) fn check_cfg_attr(attr: &Meta) -> bool {
     } else if id.is_ident("cfg") {
         if let Meta::List(ml) = attr {
             'outer: for p in ml.nested.iter() {
-                if let NestedMeta::Meta(Meta::List(ref i)) = p {
-                    if i.path.is_ident("not") {
-                        for n in i.nested.iter() {
-                            if let NestedMeta::Meta(Meta::Path(ref pth)) = n {
-                                if pth.is_ident("tarpaulin") {
-                                    ignore_span = true;
-                                    break 'outer;
+                match p {
+                    NestedMeta::Meta(Meta::List(ref i)) => {
+                        if i.path.is_ident("not") {
+                            for n in i.nested.iter() {
+                                if let NestedMeta::Meta(Meta::Path(ref pth)) = n {
+                                    if pth.is_ident("tarpaulin") {
+                                        ignore_span = true;
+                                        break 'outer;
+                                    }
                                 }
                             }
                         }
                     }
+                    NestedMeta::Meta(Meta::Path(ref i)) => {
+                        if i.is_ident("tarpaulin_skip") {
+                            ignore_span = true;
+                            break 'outer;
+                        }
+                    }
+                    _ => {}
                 }
             }
+        }
+    } else {
+        let skip_attrs = vec!["tarpaulin", "skip"];
+        let mut n = 0;
+        ignore_span = true;
+        for (segment, attr) in id.segments.iter().zip(skip_attrs.iter()) {
+            n += 1;
+            if segment.ident != attr {
+                ignore_span = false;
+            }
+        }
+        if n < skip_attrs.len() {
+            ignore_span = false;
         }
     }
     ignore_span