+++ src/source_analysis/tests.rs
@@ -720,7 +720,7 @@ fn tarpaulin_skip_attr() {
     let mut lines = LineAnalysis::new();
     let ctx = Context {
         config: &config,
-        file_contents: "#[cfg_attr(tarpaulin, skip)]
+        file_contents: "#[cfg(tarpaulin_skip)]
             fn skipped() {
                 println!(\"Hello world\");
             }
@@ -734,6 +734,11 @@ fn tarpaulin_skip_attr() {
         fn uncovered() {
             println!(\"goodbye world\");
         }
+
+        #[tarpaulin::skip]
+        fn uncovered2() {
+            println!(\"oof\");
+        }
         ",
         file: Path::new(""),
         ignore_mods: RefCell::new(HashSet::new()),
@@ -746,11 +751,13 @@ fn tarpaulin_skip_attr() {
     assert!(!lines.ignore.contains(&Lines::Line(8)));
     assert!(lines.ignore.contains(&Lines::Line(12)));
     assert!(lines.ignore.contains(&Lines::Line(13)));
+    assert!(lines.ignore.contains(&Lines::Line(17)));
+    assert!(lines.ignore.contains(&Lines::Line(18)));
 
     let mut lines = LineAnalysis::new();
     let ctx = Context {
         config: &config,
-        file_contents: "#[cfg_attr(tarpaulin, skip)]
+        file_contents: "#[cfg(tarpaulin_skip)]
         mod ignore_all {
             fn skipped() {
                 println!(\"Hello world\");